pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "flask-sqlalchemy>=3.1.1",
    "flask>=3.1.1",
    "requests>=2.32.4",
    "werkzeug>=3.1.3",
    "psycopg2-binary>=2.9.10",
    "email-validator>=2.2.0",
    "gunicorn>=23.0.0",
    "openai>=1.90.0",
]
